Generalize country filter
